try to replace uint64_t with int

+++ kernel/include/core/thread.h
@@ -63,7 +63,7 @@ typedef struct
 	 uint64_t     time_capacity;
 	 uint64_t     remaining_time_capacity;
 	 uint64_t     next_activation;
-	 uint64_t     weight;//self-adding
+	 int     weight;//self-adding
 	 pok_state_t  state;
 	 uint64_t	end_time;
 	 uint64_t     wakeup_time;
@@ -85,7 +85,7 @@ typedef struct
 	 void*        entry;            /* entrypoint of the thread  */
 	 uint64_t     period;
 	 uint64_t     deadline;
-	 uint64_t	  weight;//self-adding
+	 int	  weight;//self-adding
 	 uint64_t     time_capacity;
 	 uint32_t     stack_size;
 	 pok_state_t  state;
